requirements.txt
# FastAPI и ASGI
fastapi==0.104.0
uvicorn[standard]==0.24.0
pydantic==2.4.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.0

# Redis
redis==5.0.1

# HTTP Clients
aiohttp==3.9.0
requests==2.31.0
httpx==0.27.0

# Scraping
beautifulsoup4==4.12.2
lxml==4.9.3

# CLI
click==8.1.7
rich==13.7.0

# Utils
python-dotenv==1.0.0
python-multipart==0.0.6

ollama==0.3.0
qdrant-client==1.7.0

# Testing (опционально)
# pytest==7.4.3
# pytest-asyncio==0.21.1

# Telegram
telethon==1.35.0
pyyaml>=6.0

tqdm

langchain-core
langchain-ollama
langchain-openai
python-dotenv